requirements.txt
sseclient-py==1.7.2
requests==2.31.0
urllib3==2.0.7
python-dotenv==1.0.0
